+++ Makefile
@@ -3,3 +3,6 @@ test:
 
 format:
 	ruff format .
+
+run:
+	python -m grpc_mock
