+++ pyproject.toml
@@ -7,8 +7,10 @@ requires-python = ">=3.12"
 dependencies = [
     "bbpb>=1.4.1",
     "h2>=4.2.0",
+    "hypercorn>=0.17.3",
     "hyperframe>=6.1.0",
     "proto-schema-parser>=1.4.5",
+    "starlette>=0.45.3",
 ]
 
 [dependency-groups]
